fix: openai model to use chat completion api
